backend/requirements.txt
fastapi
uvicorn
python-multipart
openai
supabase
python-dotenv
tiktoken
pypdf
python-docx
pymupdf